Added commands and blackboard
Added a blackboard singleton to track state. It currently has hardcoded properties for selected cell / index and selection mode (ie edge, vertex or face).

Added a very basic UI for setting the selection mode. The buttons use the command pattern, but currently on set the selection state on the blackboard.